webserver/requirements.txt
blinker==1.6.2
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.5.0/de_core_news_sm-3.5.0-py3-none-any.whl
ffmpeg-python==0.2.0
filelock==3.12.2
Flask==2.3.2
future==0.18.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
llvmlite==0.40.1
mariadb==1.1.6
MarkupSafe==2.1.3
more-itertools==9.1.0
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
numba==0.57.1
numpy==1.24.4
openai-whisper==20230314
packaging==23.1
pathy==0.10.2
preshed==3.0.8
pydantic==1.10.9
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.12
thinc==8.1.10
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.6.3
urllib3==2.0.3
wasabi==1.1.2
Werkzeug==2.3.6
